Removed validation from `pos` parser and `neg` parser. Close #17.
`pos` and `neg` didn't work well with `end'` because `pos` and `neg` retuened an error if the position was at the end of the input string, while `end'` requires  it to be at the end.
Now, `pos` and `neg` don't check the position in the State so that `end'` can work.